+++ CHANGELOG.rst
@@ -15,7 +15,7 @@ Unreleased
 
 [1.27.0] - 2022-11-03
 ---------------------
-* Added XBlockSkills and XBlockSkillThrough models.
+* Added XBlockSkills and XBlockSkillData models.
 * Added related celery tasks, abstract provider, signals and commands.
 
 [1.26.0] - 2022-10-31